fix(docx-core): merge auxiliary parts in rebuild mode (#94)
External contributor @Data9210 reported that comparison output cannot be
opened in Microsoft Word when the inplace round-trip safety check fails
and the pipeline falls back to rebuild mode. The output document.xml
references comments/footnotes/endnotes whose definitions are not in the
ZIP, producing dangling references that Word refuses to open.

Root cause: `mergeAuxiliaryPartDefinitions` was gated to
`outputMode === 'inplace'`. In rebuild mode the result archive is cloned
from `originalArchive`, so any auxiliary IDs the revised side introduced
(e.g. comments newly added in the revised document) had no definitions
to bind to.

This change:

1. Hoists the auxiliary-part merge out of the inplace-only branch and
   parameterizes the merge source by mode. Inplace pulls deleted-but-
   referenced definitions from `originalArchive` (unchanged behavior),
   rebuild pulls added-but-referenced definitions from `revisedArchive`
   (the symmetric counterpart). `mergeCommentAncillaryParts` is now
   called with the same mode-aware source.

2. Preserves the conventional separator entries (`w:id="-1"` and
   `w:id="0"`) in `footnotes.xml`/`endnotes.xml` when the helper
   creates a part from scratch. The previous "remove all entries, add
   back only the missing ones" sweep would otherwise drop the
   separator and continuationSeparator entries that Word expects.

3. Adds `w:commentReference`, `w:commentRangeStart`, and
   `w:commentRangeEnd` to `LEAF_NODE_TAGS` in `atomizer.ts` so comment
   anchors survive atomization in rebuild mode. Without this, the
   reporter's specific failure (added comments) would not be observable
   even after the merge fix because the references would be stripped
   before reaching reconstruction. `w:footnoteReference` and
   `w:endnoteReference` were already in the set; this restores symmetry.

Regression coverage in
`packages/docx-core/src/integration/rebuild-auxiliary-merge.test.ts`
exercises three scenarios (added footnote, added comment, footnote on
both sides) and asserts that document.xml, the auxiliary part,
[Content_Types].xml, and document.xml.rels are all consistent after
rebuild. Test A (added footnote) is the empirically-failing case Codex
CLI reproduced during peer review. The synthetic-docx fixture builder
was extracted from the existing inplace test into
`synthetic-docx-fixture.ts` and reused.

Out of scope (filed as follow-up issues that link back to #94):
- `wrapWithDel` regex-based `<w:t>→<w:delText>` rewrite is fragile.
- `buildDocumentPreservingStructure` strips `w:commentRangeStart/End`
  outside `<w:p>`, breaking comments that span multiple paragraphs.

Fixes: #94
